# Conflicts:
#	.github/workflows/build_test.yaml
#	crane_bringup/launch/play_switcher.launch.py
#	crane_bringup/launch/play_switcher_y.launch.py
#	crane_local_planner/include/crane_local_planner/local_planner.hpp
#	crane_local_planner/src/crane_local_planner.cpp
#	session/crane_planner_base/include/crane_planner_base/planner_base.hpp
#	session/crane_planner_plugins/CMakeLists.txt
#	session/crane_planner_plugins/include/crane_planner_plugins/attacker_planner.hpp
#	session/crane_planner_plugins/include/crane_planner_plugins/ball_placement_planner.hpp
#	session/crane_planner_plugins/include/crane_planner_plugins/defender_planner.hpp
#	session/crane_planner_plugins/include/crane_planner_plugins/goalie_planner.hpp
#	session/crane_planner_plugins/include/crane_planner_plugins/kickoff_planner.hpp
#	session/crane_planner_plugins/include/crane_planner_plugins/template_planner.hpp
#	session/crane_planner_plugins/include/crane_planner_plugins/waiter_planner.hpp
#	session/crane_planner_plugins/src/attacker/attacker.cpp
#	session/crane_planner_plugins/src/ball_placement/ball_placement.cpp
#	session/crane_planner_plugins/src/ball_placement_with_skill/ball_placement_with_skill_node.cpp
#	session/crane_planner_plugins/src/goalie/goalie.cpp
#	session/crane_planner_plugins/src/kickoff/kickoff.cpp
#	session/crane_planner_plugins/src/receive/receive.cpp
#	session/crane_planner_plugins/src/template/template.cpp
#	session/crane_planner_plugins/src/waiter/waiter.cpp
#	session/crane_session_controller/config/play_situation/OUR_KICKOFF_START.yaml
#	session/crane_session_controller/include/crane_session_controller/session_module.hpp
#	session/crane_session_controller/package.xml
#	session/crane_session_controller/src/crane_session_controller.cpp
#	utility/crane_geometry/include/crane_geometry/eigen_adapter.hpp
#	utility/crane_geometry/include/crane_geometry/interval.hpp
#	utility/crane_msg_wrappers/include/crane_msg_wrappers/world_model_wrapper.hpp

+++ utility/crane_geometry/include/crane_geometry/interval.hpp
@@ -13,18 +13,18 @@
 class Interval
 {
 private:
-  std::vector<float> uppers;
-  std::vector<float> lowers;
+  std::vector<double> uppers;
+  std::vector<double> lowers;
 
 public:
   Interval() {}
 
   ~Interval() {}
 
-  void append(float a, float b)
+  void append(double a, double b)
   {
-    float upper = std::max(a, b);
-    float lower = std::min(a, b);
+    double upper = std::max(a, b);
+    double lower = std::min(a, b);
     uppers.emplace_back(upper);
     lowers.emplace_back(lower);
 
@@ -41,10 +41,10 @@ public:
     }
   }
 
-  void erase(float a, float b)
+  void erase(double a, double b)
   {
-    float upper = std::max(a, b);
-    float lower = std::min(a, b);
+    double upper = std::max(a, b);
+    double lower = std::min(a, b);
     for (size_t i = 0; i < uppers.size(); i++) {
       //完全消去
       if (uppers[i] < upper && lowers[i] > lower) {
@@ -74,22 +74,22 @@ public:
     std::sort(lowers.begin(), lowers.end());
   }
 
-  float getWidth()
+  double getWidth()
   {
-    float width = 0.f;
+    double width = 0.f;
     for (size_t i = 0; i < lowers.size(); i++) {
       width += uppers[i] - lowers[i];
     }
     return width;
   }
 
-  auto getLargestInterval() -> std::pair<float, float>
+  auto getLargestInterval() -> std::pair<double, double>
   {
-    float max_width = 0.f;
-    float max_lower = 0.f;
-    float max_upper = 0.f;
+    double max_width = 0.f;
+    double max_lower = 0.f;
+    double max_upper = 0.f;
     for (size_t i = 0; i < lowers.size(); i++) {
-      float width = uppers[i] - lowers[i];
+      double width = uppers[i] - lowers[i];
       if (width > max_width) {
         max_width = width;
         max_lower = lowers[i];
